LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window-expected.txt


FAIL page-created Error (worker) The object can not be cloned.
FAIL page-created Error (cross-site iframe) The object can not be cloned.
PASS page-created DOMException (worker)
PASS page-created DOMException (cross-site iframe)
FAIL JS-engine-created TypeError (worker) The object can not be cloned.
FAIL JS-engine-created TypeError (cross-site iframe) The object can not be cloned.
FAIL web API-created TypeError (worker) The object can not be cloned.
FAIL web API-created TypeError (cross-site iframe) The object can not be cloned.
FAIL web API-created DOMException (worker) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:41:31\n@http://localhost:8800/resources/testharness.js:2087:30\nasync_test@http://localhost:8800/resources/testharness.js:594:38\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:40:13\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
FAIL web API-created DOMException (cross-site iframe) assert_equals: expected (string) "createElement@[native code]\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:33:27\n@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:60:31\n@http://localhost:8800/resources/testharness.js:2087:30\nasync_test@http://localhost:8800/resources/testharness.js:594:38\nstackTests@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:57:13\nglobal code@http://localhost:8800/html/infrastructure/safe-passing-of-structured-data/structured-cloning-error-stack-optional.sub.window.js:31:11" but got (undefined) undefined
